NIFI-13236 Moved logoutSupported from Configuration to Current User
